Update strncpy.c
Actually, it should be type size_t, since the type is already available in this context (max).

+++ sources/strncpy.c
@@ -10,7 +10,7 @@
 char *strncpy(char *dst, const char *src, size_t max)
 {
 	char *dscan;
-	unsigned long count;
+	size_t count;
 
 	dscan = dst;
 	count = max;
